🧹 refactor: Extract OAuthButton component from AuthCard
The inline provider button rendering in AuthCard branched on the presence
of a logo glyph: branded providers needed children for the logo+text
composition (and an eslint-disable for click-ui's button-requires-label
rule), while unbranded providers used the label prop directly. Both paths
shared identical type, disabled, onClick, and fillWidth wiring.

Moving the branch into a dedicated OAuthButton scopes the eslint-disable
to a single helper, collapses the provider-rendering site in AuthCard to
a plain map, and gives the future github/discord/saml additions a single
file to land their glyph rules in.
